apps/worker/requirements.txt
pydantic==2.8.2
python-dotenv==1.0.1
langchain==0.2.12
langchain-community==0.2.11
qdrant-client==1.11.3
pypdf==4.2.0
markdown==3.6
httpx==0.27.0

beautifulsoup4==4.12.3
lxml[html_clean]==5.3.0
readability-lxml==0.8.1
tldextract==5.1.2

numpy==1.26.4
